5.SpatialNullmodels.r
Estimates spatial null models for mean family age (MAF) across a global grid of 2x2 (resolution as in the original estimation of the presence absence matrix). The estimation of the null models is performed by shuffling entries in the PAM in two ways:
- NULL 1 randomly shuffles occurrences within sites (i.e., per-row shuffling), keeping the observed species richness.
- NULL 2 randomly shuffles occurrences within families (i.e., per-column shuffling), keeping the observed family prevalence.
Continues with R objects generated in previous steps (loads objects, saves and loads objects for plotting).